pyproject.toml
[project]
name = "fintech-app"
version = "0.1.0"
description = "Production-style FastAPI fintech application consuming a custom Elixir RDBMS"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "pydantic",
    "python-jose",
]

[project.optional-dependencies]
dev = ["python-dotenv"]


[tool.pytest.ini_options]
testpaths = ["tests"]
